Reword bit about when to not squash PRs
Hopefully this makes it less of a decision you need to make (the point of this section was only to tell people to “squash & merge” instead of “merge”). If you don’t know the difference, you shouldn’t need to worry about it. Squash & merge unless you have a specific reason not to do so.